v1_archive/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.2.0
python-dotenv>=1.0.0

# Video processing
moviepy>=2.0.0
ffmpeg-python>=0.2.0

# Audio transcription
assemblyai>=0.28.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Machine learning (optional)
scikit-learn>=1.3.0

# Utilities
pathlib2>=2.3.7; python_version < '3.4'
